Rename Details component to HolidayDetails component
