+++ src/App.js
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Home from './components/pages/Home';
-import Details from './components/pages/Details';
+import Details from './components/pages/HolidaysDetails';
 
 function App() {
   return (
